% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sf_to_rast}
\alias{sf_to_rast}
\title{Convert an sf points data frame grid to a SpatRaster object}
\usage{
sf_to_rast(x, digits = 6)
}
\arguments{
\item{x}{An sf points data frame with locations on a grid}

\item{digits}{The number of significant digits in the coordinate values for snapping to a grid}
}
\description{
Convert an sf points data frame grid to a SpatRaster object
}
\author{
Devin S. Johnson
}
